RAC-102 refactor: res 패키지로 이동
