+++ src/main/java/com/postgraduate/domain/mentoring/application/usecase/MentoringInfoUseCase.java
@@ -1,7 +1,7 @@
 package com.postgraduate.domain.mentoring.application.usecase;
 
-import com.postgraduate.domain.mentoring.application.dto.AppliedMentoringDetailResponse;
-import com.postgraduate.domain.mentoring.application.dto.AppliedMentoringResponse;
+import com.postgraduate.domain.mentoring.application.dto.res.AppliedMentoringDetailResponse;
+import com.postgraduate.domain.mentoring.application.dto.res.AppliedMentoringResponse;
 import com.postgraduate.domain.mentoring.application.dto.AppliedMentoringInfo;
 import com.postgraduate.domain.mentoring.application.mapper.MentoringMapper;
 import com.postgraduate.domain.mentoring.domain.entity.Mentoring;
